Big changes to the shop. Added rotate, zoom, and secondary tag systems.
Added the ability to rotate the preview mesh by click-and-drag on the preview mesh (white squares are visibile temporarily for testing purposes).
Added the ability to zoom in on the cosmetic item (does not move the camera, just scales up the preview mesh).
Added the ability to filter cosmetics by type (to compliment the existing rarity system).